requirements.txt
# requirements.txt for SEO Content Generator with OpenAI Project

streamlit==1.33.0
langchain-openai==0.1.0
python-dotenv==1.0.0
langchain-core==0.1.0